Rearchitect bridge to pay fee before server interacts with target network
This eliminates the possibility for the user to hack the frontend to manipulate the fee paid by storing the transaction in the backend and communicating the fee for this transaction to the target network bridge contract where a check is made that the user paid a sufficient fee.

This change also reduces the complexity of the server setup by only using a single server instead of two. There will be a more stringent requirement on the server to be secure, or other protections in place.